feat: new visual for forms register
