+++ handler/catalog_fornecedores.go
@@ -2,6 +2,7 @@ package handler
 
 import (
 	"net/http"
+	"strconv"
 
 	"github.com/gin-gonic/gin"
 	"github.com/lcardelli/fornecedores/schemas"
@@ -34,9 +35,20 @@ func CatalogFornecedoresHandler(c *gin.Context) {
 
 	// Buscar serviços para o filtro
 	var services []schemas.Service
-	if err := db.Find(&services).Error; err != nil {
-		c.HTML(http.StatusInternalServerError, "error.html", gin.H{"error": "Erro ao buscar serviços"})
-		return
+	if categoryID != "" {
+		categoryIDInt, _ := strconv.Atoi(categoryID)
+		if err := db.Joins("JOIN supplier_services ON services.id = supplier_services.service_id").
+			Joins("JOIN supplier_links ON supplier_links.id = supplier_services.supplier_link_id").
+			Where("supplier_links.category_id = ?", categoryIDInt).
+			Distinct().Find(&services).Error; err != nil {
+			c.HTML(http.StatusInternalServerError, "error.html", gin.H{"error": "Erro ao buscar serviços"})
+			return
+		}
+	} else {
+		if err := db.Find(&services).Error; err != nil {
+			c.HTML(http.StatusInternalServerError, "error.html", gin.H{"error": "Erro ao buscar serviços"})
+			return
+		}
 	}
 
 	// Construir a query para SupplierLinks
@@ -90,4 +102,21 @@ func CatalogFornecedoresHandler(c *gin.Context) {
 			"activeMenu": "catalogo",
 		},
 	})
+}
+
+// Adicione esta nova função para lidar com a requisição AJAX
+func GetServicesByCategoryHandler(c *gin.Context) {
+	categoryID := c.Param("categoryId")
+	categoryIDInt, _ := strconv.Atoi(categoryID)
+
+	var services []schemas.Service
+	if err := db.Joins("JOIN supplier_services ON services.id = supplier_services.service_id").
+		Joins("JOIN supplier_links ON supplier_links.id = supplier_services.supplier_link_id").
+		Where("supplier_links.category_id = ?", categoryIDInt).
+		Distinct().Find(&services).Error; err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao buscar serviços"})
+		return
+	}
+
+	c.JSON(http.StatusOK, services)
 }
